extractapp/requirements.txt
langextract
langextract-litellm
litellm
python-dotenv